Positioning, Fonts and images
Added a reference link to a number of google fonts. Added and commented the font family in the style sheet. Need to figure out a better way of saving that.

Created a middle and bottom container that has a profile picture and skill information. Also included a link to my new email address and LinkedIn page.